# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShelfOps API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/recurrence-suggestions/{sourceSuggestionId}/decision": {
      "post": {
        "summary": "Confirm, dismiss, or correct a recurrence suggestion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "state",
                  "expectedVersion",
                  "idempotencyKey"
                ],
                "properties": {
                  "state": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "confirmed"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "dismissed"
                        ]
                      }
                    ]
                  },
                  "expectedVersion": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "idempotencyKey": {
                    "minLength": 1,
                    "pattern": "\\S",
                    "type": "string"
                  },
                  "note": {
                    "minLength": 1,
                    "pattern": "\\S",
                    "type": "string"
                  },
                  "correctionReason": {
                    "minLength": 1,
                    "pattern": "\\S",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "sourceSuggestionId",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "sourceSuggestionId",
                    "decisionId",
                    "state",
                    "sequence",
                    "incidentId",
                    "version",
                    "decidedAt",
                    "correlationId"
                  ],
                  "properties": {
                    "sourceSuggestionId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "decisionId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "state": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "confirmed"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "dismissed"
                          ]
                        }
                      ]
                    },
                    "sequence": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "predecessorSequence": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "incidentId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "version": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "decidedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId",
                    "fields"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation-failed"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "name",
                          "code"
                        ],
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "authentication-required"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "forbidden"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "not-found"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "idempotency-conflict"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "description": "Server-generated identifier for support correlation.",
                          "type": "string"
                        }
                      }
                    },
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "stale-version"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "description": "Server-generated identifier for support correlation.",
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Correlation for this HTTP request; replays receive a new value."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "temporarily-unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shelfops_session"
      }
    },
...
  }
}